pyproject.toml
[project]
name = "contextual-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "contextual-client>=0.7.0",
    "ipykernel>=6.29.5",
    "keybert>=0.9.0",
    "langchain-openai>=0.3.18",
    "langchain>=0.3.25",
    "openai>=1.82.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "langchain-community>=0.3.24",
    "pypdf>=5.5.0",
    "pymupdf>=1.26.0",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "tabulate>=0.9.0",
    "seaborn>=0.13.2",
]
